@@ -15,7 +15,8 @@
     "@types/node": "^8.0.47",
     "awesome-typescript-loader": "^3.2.3",
     "source-map-loader": "^0.2.3",
-    "typescript": "^2.5.3",
+    "ts-loader": "^3.1.1",
+    "typescript": "^2.6.1",
     "webpack-dev-server": "^2.9.3"
   },
   "dependencies": {
